pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.0.0",
    "aiofiles>=23.0.0",
    "python-dateutil>=2.8.0",
    "itsdangerous>=2.0.0",
    "email-validator>=2.0.0"
]

[project.scripts]
lab2 = "app.main:run"
